Add wall bounciness
Colliding with the wall without punishment, besides bouncing, does not
feel good.

Bounciness doesn't stop placing from going into the wall, since it can
sometimes be used to their advantage.

However, adding a HP / damage factor can be very interesting. It can
lead to tense moments when you have to be very careful about colliding
with the walls or other ships, because you are about to break down. And
it can lead to moments where you can trade some HP for a small advantage
that colliding with the wall gives. Or even to brake someone's ship at
the cost of some of your HP, hehe.